control: rename 2 packages have -cisco suffixes
This entails updating a bunch of places where the package name
"openvswitch-switch" was used, and also adding in debian/control
fields to say "Even though my package name is
openvswitch-switch-cisco, you can pretend that I'm openvswitch-switch
version 2.9.8-0ubuntu0.18.04.2".  This enables
openvswitch-switch-cisco to serve as a valid dependency for anything
that requires openvswitch-switch.

Ditto for openvswitch-common (which became openvswitch-common-cisco).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>